Feito calculo de juros

+++ src/TesteTecnicoBancos.Communication/Requests/RequestRegisterBankJson.cs
@@ -3,5 +3,5 @@ public class RequestRegisterBankJson
 {
     public string Name { get; set; } = string.Empty;
     public int Code { get; set; } 
-    public float Interest { get; set; }
+    public decimal Interest { get; set; }
 }
