+++ src/TesteTecnicoBancos.Communication/Requests/RequestRegisterBoletoJson.cs
@@ -5,7 +5,7 @@ public class RequestRegisterBoletoJson
     public string PayeeCpfCnpj { get; set; } = string.Empty;
     public string RecipientName { get; set; } = string.Empty;
     public string RecipientCpfCnpj { get; set; } = string.Empty;
-    public float Value { get; set; }
+    public decimal Value { get; set; }
     public DateTime DueDate { get; set; }
     public string Obs { get; set; } = string.Empty;
     public long BankId { get; set; }
