+++ src/TesteTecnicoBancos.Communication/Responses/ResponseBankJson.cs
@@ -4,5 +4,5 @@ public class ResponseBankJson
     public long Id { get; set; }
     public string? Name { get; set; }
     public int Code { get; set; }
-    public float Interest { get; set; }
+    public decimal Interest { get; set; }
 }
