+++ src/TesteTecnicoBancos.Communication/Responses/ResponseRegisteredBankJson.cs
@@ -3,5 +3,5 @@ public class ResponseRegisteredBankJson
 {
     public string Name { get; set; } = string.Empty;
     public int Code { get; set; }
-    public float Interest { get; set; }
+    public decimal Interest { get; set; }
 }
