+++ src/TesteTecnicoBancos.Communication/Responses/ResponseRegisteredBoletoJson.cs
@@ -5,6 +5,6 @@ public class ResponseRegisteredBoletoJson
     public string PayeeCpfCnpj { get; set; } = string.Empty;
     public string RecipientName { get; set; } = string.Empty;
     public string RecipientCpfCnpj { get; set; } = string.Empty;
-    public float Value { get; set; }
+    public decimal Value { get; set; }
     public DateTime Duedate { get; set; }
 }
